+++ src/components/widgets/EntityInfoWidget/EntityInfoWidget.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { EuiCard, EuiFlexItem, EuiLoadingSpinner, EuiSpacer, EuiText } from "@elastic/eui";
 import { OlsApi } from "../../../api/OlsApi";
 import { useQuery } from 'react-query'
-import {getErrorMessageToDisplay} from "../index";
+import { getErrorMessageToDisplay } from "../../../utils/helper";
 
 export interface EntityInfoWidgetProps {
     api: string;
