+++ src/components/widgets/MetadataWidget/TabWidget/CrossRefWidget/CrossRefTabWidget.tsx
@@ -8,8 +8,7 @@ import {
 } from "@elastic/eui";
 import { OlsApi } from '../../../../../api/OlsApi'
 import { useQuery } from 'react-query'
-import {getPreferredOntologyJSON} from "../../index";
-import {getErrorMessageToDisplay} from "../../../index";
+import { getErrorMessageToDisplay, getPreferredOntologyJSON } from "../../../../../utils/helper";
 
 export interface CrossRefWidgetProps {
   iri: string;
