+++ src/components/widgets/MetadataWidget/TabWidget/TabWidget.tsx
@@ -8,7 +8,7 @@ import { CrossRefTabWidget } from "./CrossRefWidget";
 import { HierarchyWidget } from "./HierarchyWidget";
 import {useQuery} from "react-query";
 import {OlsApi} from "../../../../api/OlsApi";
-import {getPreferredOntologyJSON} from "../index";
+import { getPreferredOntologyJSON } from "../../../../utils/helper";
 
 export interface TabWidgetProps {
   iri: string;
