+++ src/components/widgets/MetadataWidget/TitleWidget/TitleWidget.tsx
@@ -2,8 +2,7 @@ import React from "react";
 import {useQuery} from "react-query";
 import {EuiLoadingSpinner, EuiText} from "@elastic/eui";
 import {OlsApi} from "../../../../api/OlsApi";
-import {getPreferredOntologyJSON} from "../index";
-import {getErrorMessageToDisplay} from "../../index";
+import { getErrorMessageToDisplay, getPreferredOntologyJSON } from "../../../../utils/helper";
 
 export interface TitleWidgetProps {
     iri?: string;
