+++ src/components/widgets/index.ts
@@ -7,11 +7,3 @@ export * from "./OntologyInfoWidget";
 export * from "./SearchBarWidget";
 export * from "./SearchResultsListWidget";
 export * from "./EntityInfoWidget";
-
-export function getErrorMessageToDisplay(error: any, messagePlaceholder = "information"): string {
-    const error_msg : string = error.message;
-    if(error_msg === ("Response contains 0 elements")) {
-        return "No elements found";
-    }
-    else return `No ${messagePlaceholder} available`;
-}
